+++ internal/payload/embed_windows.go
@@ -6,3 +6,6 @@ import _ "embed"
 
 //go:embed ls-wrapper.exe
 var LSWrapper []byte
+
+//go:embed devin-wrapper.exe
+var DevinWrapper []byte